rs_tfhe::bit_utils

Trait AsBits

pub trait AsBits<T> {
    // Required methods
    fn to_bits(self) -> Vec<bool>;
    fn encrypt(self, key: &SecretKey) -> Vec<Ciphertext> ;
}

Required Methods

fn to_bits(self) -> Vec<bool>

Represent the bits of the byte as an array of boolean values (bits). Array is in big-endian order, where MSB is the first value of the array.

fn encrypt(self, key: &SecretKey) -> Vec<Ciphertext>

Encrypt the bits of the type as an array of LweSample values (cipherbits).